# api/generated/models.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package generated

import (
	"time"
)

// Defines values for ErrorResponseErrorCode.
const (
	INTERNALERROR    ErrorResponseErrorCode = "INTERNAL_ERROR"
	NOCANDIDATE      ErrorResponseErrorCode = "NO_CANDIDATE"
	NOTASSIGNED      ErrorResponseErrorCode = "NOT_ASSIGNED"
	PREXISTS         ErrorResponseErrorCode = "PR_EXISTS"
	PRMERGED         ErrorResponseErrorCode = "PR_MERGED"
	PRNOTFOUND       ErrorResponseErrorCode = "PR_NOT_FOUND"
	PRNOTOPEN        ErrorResponseErrorCode = "PR_NOT_OPEN"
	REVIEWERNOTFOUND ErrorResponseErrorCode = "REVIEWER_NOT_FOUND"
	TEAMEXISTS       ErrorResponseErrorCode = "TEAM_EXISTS"
	TEAMNOTFOUND     ErrorResponseErrorCode = "TEAM_NOT_FOUND"
	USERINACTIVE     ErrorResponseErrorCode = "USER_INACTIVE"
	USERNOTFOUND     ErrorResponseErrorCode = "USER_NOT_FOUND"
	VALIDATIONERROR  ErrorResponseErrorCode = "VALIDATION_ERROR"
)

// Defines values for PullRequestStatus.
const (
	PullRequestStatusMERGED PullRequestStatus = "MERGED"
	PullRequestStatusOPEN   PullRequestStatus = "OPEN"
)

// Defines values for PullRequestShortStatus.
const (
	PullRequestShortStatusMERGED PullRequestShortStatus = "MERGED"
	PullRequestShortStatusOPEN   PullRequestShortStatus = "OPEN"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error struct {
		Code    ErrorResponseErrorCode `json:"code"`
		Message string                 `json:"message"`
	} `json:"error"`
}

// ErrorResponseErrorCode defines model for ErrorResponse.Error.Code.
type ErrorResponseErrorCode string

// PullRequest defines model for PullRequest.
type PullRequest struct {
	// AssignedReviewers user_id назначенных ревьюверов (0..2)
	AssignedReviewers []string          `json:"assigned_reviewers"`
	AuthorId          string            `json:"author_id"`
	CreatedAt         *time.Time        `json:"createdAt"`
	MergedAt          *time.Time        `json:"mergedAt"`
	PullRequestId     string            `json:"pull_request_id"`
	PullRequestName   string            `json:"pull_request_name"`
	Status            PullRequestStatus `json:"status"`
}

// PullRequestStatus defines model for PullRequest.Status.
type PullRequestStatus string

// PullRequestShort defines model for PullRequestShort.
type PullRequestShort struct {
	AuthorId        string                 `json:"author_id"`
	PullRequestId   string                 `json:"pull_request_id"`
	PullRequestName string                 `json:"pull_request_name"`
	Status          PullRequestShortStatus `json:"status"`
}

// PullRequestShortStatus defines model for PullRequestShort.Status.
type PullRequestShortStatus string

// Team defines model for Team.
type Team struct {
	Members  []TeamMember `json:"members"`
	TeamName string       `json:"team_name"`
}

// TeamMember defines model for TeamMember.
type TeamMember struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// User defines model for User.
type User struct {
	IsActive bool   `json:"is_active"`
	TeamName string `json:"team_name"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// TeamNameQuery defines model for TeamNameQuery.
type TeamNameQuery = string

// UserIdQuery defines model for UserIdQuery.
type UserIdQuery = string

// CreatePullRequestJSONBody defines parameters for CreatePullRequest.
type CreatePullRequestJSONBody struct {
	AuthorId        string `json:"author_id"`
	PullRequestId   string `json:"pull_request_id"`
	PullRequestName string `json:"pull_request_name"`
}

// MergePullRequestJSONBody defines parameters for MergePullRequest.
type MergePullRequestJSONBody struct {
	PullRequestId string `json:"pull_request_id"`
}

// ReassignReviewerJSONBody defines parameters for ReassignReviewer.
type ReassignReviewerJSONBody struct {
	OldUserId     string `json:"old_user_id"`
	PullRequestId string `json:"pull_request_id"`
}

// GetTeamParams defines parameters for GetTeam.
type GetTeamParams struct {
	// TeamName Уникальное имя команды
	TeamName TeamNameQuery `form:"team_name" json:"team_name"`
}

// GetUserAssignedReviewsParams defines parameters for GetUserAssignedReviews.
type GetUserAssignedReviewsParams struct {
	// UserId Идентификатор пользователя
	UserId UserIdQuery `form:"user_id" json:"user_id"`
}

// SetUserActiveJSONBody defines parameters for SetUserActive.
type SetUserActiveJSONBody struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
}

// CreatePullRequestJSONRequestBody defines body for CreatePullRequest for application/json ContentType.
type CreatePullRequestJSONRequestBody CreatePullRequestJSONBody

// MergePullRequestJSONRequestBody defines body for MergePullRequest for application/json ContentType.
type MergePullRequestJSONRequestBody MergePullRequestJSONBody

// ReassignReviewerJSONRequestBody defines body for ReassignReviewer for application/json ContentType.
type ReassignReviewerJSONRequestBody ReassignReviewerJSONBody

// AddTeamJSONRequestBody defines body for AddTeam for application/json ContentType.
type AddTeamJSONRequestBody = Team

// SetUserActiveJSONRequestBody defines body for SetUserActive for application/json ContentType.
type SetUserActiveJSONRequestBody SetUserActiveJSONBody